reorganized package; bumped version
