+++ docs/index.rst
@@ -1,9 +1,9 @@
-Welcome to Kutana!
-==================
+Kutana
+======
 
 .. image:: _static/kutana-logo-512.png
     :alt: Kutana
-    :scale: 65%
+    :scale: 60%
 
 The engine for developing bots for soc. networks,
 instant messengers and other systems.
