+++ docs/src/kutana.rst
@@ -7,7 +7,6 @@ Subpackages
 .. toctree::
 
     kutana.controllers
-    kutana.tools
 
 Submodules
 ----------
@@ -24,6 +23,14 @@ kutana.executor module
 ----------------------
 
 .. automodule:: kutana.executor
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+kutana.functions module
+-----------------------
+
+.. automodule:: kutana.functions
     :members:
     :undoc-members:
     :show-inheritance:
@@ -52,10 +59,10 @@ kutana.plugin module
     :undoc-members:
     :show-inheritance:
 
-kutana.plugindata module
+kutana.structures module
 ------------------------
 
-.. automodule:: kutana.plugindata
+.. automodule:: kutana.structures
     :members:
     :undoc-members:
     :show-inheritance:
