--- docs/src/kutana.tools.rst
@@ -1,30 +1,0 @@
-kutana.tools package
-====================
-
-Submodules
-----------
-
-kutana.tools.functions module
------------------------------
-
-.. automodule:: kutana.tools.functions
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kutana.tools.structures module
-------------------------------
-
-.. automodule:: kutana.tools.structures
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kutana.tools
-    :members:
-    :undoc-members:
-    :show-inheritance:
