+++ kutana/__init__.py
@@ -7,7 +7,7 @@ from kutana.executor import *  # lgtm [py/polluting-import]
 from kutana.controllers.debug import *  # lgtm [py/polluting-import]
 from kutana.controllers.vk import *  # lgtm [py/polluting-import]
 
-from kutana.tools.structures import *  # lgtm [py/polluting-import]
-from kutana.tools.functions import *  # lgtm [py/polluting-import]
+from kutana.structures import *  # lgtm [py/polluting-import]
+from kutana.functions import *  # lgtm [py/polluting-import]
 
 name = "kutana"
