+++ setup.py
@@ -10,7 +10,7 @@ import os
 import sys
 
 
-VERSION = "0.3.2"
+VERSION = "0.4.0"
 
 
 class VerifyVersionCommand(install):
